+++ lib/assets.dart
@@ -1,0 +1,1 @@
+export 'package:dracone/assets/image_asset.dart';