+++ lib/components.dart
@@ -4,4 +4,5 @@ export 'package:dracone/components/local_to_world.dart';
 export 'package:dracone/components/origin.dart';
 export 'package:dracone/components/parent.dart';
 export 'package:dracone/components/rectangle.dart';
+export 'package:dracone/components/sprite.dart';
 export 'package:dracone/components/transform.dart';