+++ lib/components/rectangle.dart
@@ -4,18 +4,18 @@ import 'package:flutter/material.dart';
 
 class Rectangle with Component<Rectangle> {
   Size size;
-  Color color;
+  Paint paint;
 
-  Rectangle({@required this.size, @required this.color});
+  Rectangle({@required this.size, @required this.paint});
 
   @override
   Rectangle copy() {
-    return Rectangle(size: size.copy(), color: color);
+    return Rectangle(size: size.copy(), paint: paint);
   }
 
   @override
-  Rectangle copyWith({Size size, Color color}) {
+  Rectangle copyWith({Size size, Paint color}) {
     return Rectangle(
-        size: size ?? this.size.copy(), color: color ?? this.color);
+        size: size ?? this.size.copy(), paint: color ?? this.paint);
   }
 }