+++ lib/resources.dart
@@ -1,3 +1,4 @@
 export 'package:dracone/resources/dracone_context.dart';
+export 'package:dracone/resources/data_cache.dart';
 export 'package:dracone/resources/renderer.dart';
 export 'package:dracone/resources/screen_dimensions.dart';