+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## [0.0.2] - Sprite rendering, Assets system
+
+- Sprite rendering
+- Assets system
+- Assets data cache
+
 ## [0.0.1] - Initial release
 
 - Initial release